pyproject.toml
[tool.poetry]
name = "lowcode-platform-datascience"
version = "0.1.0"
description = ""
authors = ["Haohanblue <haohanblue@foxmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
pandas = "^2.2.3"
pytesseract = "^0.3.13"
pillow = "^11.1.0"
python-multipart = "^0.0.20"
matplotlib = "^3.10.0"
cnocr = {extras = ["serve"], version = "^2.3.1"}
onnxruntime = "^1.20.1"
cnstd = "^1.2.5.2"
onnx = "^1.17.0"
rapidocr-onnxruntime = "1.3.25"
huggingface-hub = "^0.27.0"
zhipuai = "^2.1.5.20250106"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
